Remove Check(XmlReader)
The process was holding an open descriptor to the XSD file during
checking, which is wasteful (and problematic if fixing issues while
stopped at a breakpoint).  This was due to the unclear ownership
semantics of Check(XmlReader) since it probably should allow the caller
to Close/Dispose the XmlReader at their discretion but it doesn't return
until checks are done.  Remove it.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>